+++ src/main/java/de/vette/idea/neos/lang/eel/completion/EelContextCompletionProvider.java
@@ -18,7 +18,7 @@ import org.jetbrains.annotations.NotNull;
 import java.util.Collection;
 import java.util.List;
 
-public class EelCompletionProvider extends CompletionProvider<CompletionParameters> {
+public class EelContextCompletionProvider extends CompletionProvider<CompletionParameters> {
 
     @Override
     protected void addCompletions(@NotNull CompletionParameters parameters, @NotNull ProcessingContext context, @NotNull CompletionResultSet result) {