+++ src/main/java/de/vette/idea/neos/lang/fusion/psi/impl/ext/FusionPathImplMixin.java
@@ -61,5 +61,4 @@ public abstract class FusionPathImplMixin extends FusionElementImpl implements F
 
         return this.getText();
     }
-
 }