+++ src/main/java/de/vette/idea/neos/lang/fusion/stubs/FusionFileStub.java
@@ -36,7 +36,7 @@ public class FusionFileStub extends PsiFileStubImpl<FusionFile> {
 
         @Override
         public int getStubVersion() {
-            return 5;
+            return 6;
         }
 
         @NotNull